Conflicts:
	GameEngine/src/com/javagameengine/TestGame.java
	GameEngine/src/com/javagameengine/math/Color4f.java

+++ GameEngine/src/com/javagameengine/scene/component/CameraStatic.java
@@ -63,7 +63,7 @@ public class CameraStatic extends Component implements Listener
 			s.getEventManager().registerListener(this);
 	}
 
-	public void onUpdate(int delta)
+	public void onUpdate(float delta)
 	{
 	}
 }
